Fix issue: Tooltip - Workspaces tooltip appears above Workspaces row and overlaps with Workspaces header
